+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## CHANGELOG
 
+1 April:
+- Added back the other routes to demonstrate React Router
+
 30 March:
 - Initial commit